feat(frontend): move the Settings detail editors into Edit-profile/Edit-hours modals
The Settings page led with the compact 2-column control panel the design
specifies, but then stacked the full-width OrgSection (User profile, Address,
Professional details, an inline availability editor) and SecuritySection below
it. On a wide screen each read row pinned its label to the far left and value
to the far right, leaving a large empty gap - the "too much space, bad
structure" QA flagged.

The design (PIMS - Organization & Settings) keeps the Settings screen to the
panel alone and reaches the editors from the Personal card's affordances: an
"Edit profile" surface and a dedicated centered "Edit hours" availability
modal. Match it:

- Split OrgSection into ProfileDetails (the three profile cards + their save
  handlers) and fold the availability state/loading/save into a new
  HoursEditModal; delete OrgSection.
- Add ProfileEditModal (identity, address, professional details, plus the
  account security controls) and HoursEditModal (single-column availability
  with a Cancel/Save footer that closes on success), both on the shared
  centered Modal primitive - which already traps focus, restores on close,
  re-forms to a bottom sheet on phones, and whitelists the ProfileCard
  dropdowns/datepicker so they never dismiss it.
- Personal opens the two modals via onEditProfile/onEditHours props; the page
  body is now just the panel. ProfileCard stays shared and untouched.

Net -998/+79 lines. Coverage: index/Personal/ProfileDetails/ProfileEditModal
100%, HoursEditModal 100% stmts.
