+++ apps/frontend/src/app/__tests__/pages/Settings/index.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import Settings from '@/app/features/settings/pages/Settings';
 
@@ -25,13 +25,22 @@ jest.mock('next/dynamic', () => ({
         return <div>Appearance Preference</div>;
       }
 
-      if (source.includes('Sections/OrgSection')) {
-        const MockOrgSection = (
-          jest.requireMock('@/app/features/settings/pages/Settings/Sections/OrgSection') as {
+      if (source.includes('Sections/ProfileEditModal')) {
+        const MockProfileEditModal = (
+          jest.requireMock('@/app/features/settings/pages/Settings/Sections/ProfileEditModal') as {
             default: React.FC<Record<string, unknown>>;
           }
         ).default;
-        return <MockOrgSection {...props} />;
+        return <MockProfileEditModal {...props} />;
+      }
+
+      if (source.includes('Sections/HoursEditModal')) {
+        const MockHoursEditModal = (
+          jest.requireMock('@/app/features/settings/pages/Settings/Sections/HoursEditModal') as {
+            default: React.FC<Record<string, unknown>>;
+          }
+        ).default;
+        return <MockHoursEditModal {...props} />;
       }
 
       if (source.includes('Sections/TimezonePreference')) {
@@ -51,10 +60,6 @@ jest.mock('next/dynamic', () => ({
           }
         ).default;
         return <MockCompanionTerminologyPreference {...props} />;
-      }
-
-      if (source.includes('Sections/SecuritySection')) {
-        return <div>Security Section</div>;
       }
 
       if (source.includes('Sections/DeleteProfile')) {
@@ -81,12 +86,27 @@ jest.mock('@/app/ui/layout/guards/ProtectedRoute', () => ({
 
 jest.mock('@/app/features/settings/pages/Settings/Sections/Personal', () => ({
   __esModule: true,
-  default: () => <div>Personal Card</div>,
+  default: ({ onEditProfile, onEditHours }: any) => (
+    <div>
+      <span>Personal Card</span>
+      <button type="button" onClick={onEditProfile}>
+        open-profile
+      </button>
+      <button type="button" onClick={onEditHours}>
+        open-hours
+      </button>
+    </div>
+  ),
 }));
 
-jest.mock('@/app/features/settings/pages/Settings/Sections/OrgSection', () => ({
+jest.mock('@/app/features/settings/pages/Settings/Sections/ProfileEditModal', () => ({
   __esModule: true,
-  default: () => <div>Org Section</div>,
+  default: ({ showModal }: any) => <div>{`Profile Modal ${showModal ? 'open' : 'closed'}`}</div>,
+}));
+
+jest.mock('@/app/features/settings/pages/Settings/Sections/HoursEditModal', () => ({
+  __esModule: true,
+  default: ({ showModal }: any) => <div>{`Hours Modal ${showModal ? 'open' : 'closed'}`}</div>,
 }));
 
 jest.mock('@/app/features/settings/pages/Settings/Sections/DeleteProfile', () => ({
@@ -108,15 +128,26 @@ describe('Settings page', () => {
     expect(screen.getByText('Changes save automatically')).toBeInTheDocument();
   });
 
-  it('renders settings sections inside protected route', () => {
+  it('renders the compact panel and keeps the editor modals closed', () => {
     render(<Settings />);
 
     expect(screen.getByTestId('protected')).toBeInTheDocument();
     expect(screen.getByText('Personal Card')).toBeInTheDocument();
-    expect(screen.getByText('Org Section')).toBeInTheDocument();
     expect(screen.getByText('Companion Terminology')).toBeInTheDocument();
     expect(screen.getByText('Appearance Preference')).toBeInTheDocument();
-    expect(screen.getByText('Security Section')).toBeInTheDocument();
     expect(screen.getByText('Delete Profile')).toBeInTheDocument();
+    // The detailed editors are modals, closed until their affordance is used.
+    expect(screen.getByText('Profile Modal closed')).toBeInTheDocument();
+    expect(screen.getByText('Hours Modal closed')).toBeInTheDocument();
+  });
+
+  it('opens the profile and hours modals from the Personal card affordances', () => {
+    render(<Settings />);
+
+    fireEvent.click(screen.getByRole('button', { name: 'open-profile' }));
+    expect(screen.getByText('Profile Modal open')).toBeInTheDocument();
+
+    fireEvent.click(screen.getByRole('button', { name: 'open-hours' }));
+    expect(screen.getByText('Hours Modal open')).toBeInTheDocument();
   });
 });
