--- apps/frontend/src/app/features/settings/pages/Settings/Sections/OrgSection.tsx
@@ -1,459 +1,0 @@
-import React, { useEffect, useMemo, useState, startTransition } from 'react';
-import { IoGlobeOutline } from 'react-icons/io5';
-import ProfileCard from '@/app/features/organization/pages/Organization/Sections/ProfileCard';
-import Availability from '@/app/features/appointments/components/Availability/Availability';
-import { usePrimaryOrgWithMembership } from '@/app/hooks/useOrgSelectors';
-import { Primary } from '@/app/ui/primitives/Buttons';
-import {
-  AvailabilityState,
-  convertAvailability,
-  convertFromGetApi,
-  daysOfWeek,
-  DEFAULT_INTERVAL,
-  hasAtLeastOneAvailability,
-} from '@/app/features/appointments/components/Availability/utils';
-import {
-  upsertAvailability,
-  upsertTeamAvailability,
-} from '@/app/features/organization/services/availabilityService';
-import { usePrimaryAvailability } from '@/app/hooks/useAvailabiities';
-import { usePrimaryOrgProfile } from '@/app/hooks/useProfiles';
-import {
-  Gender,
-  UserProfile,
-  UserEmploymentTypeOptions,
-  UserGenderOptions,
-} from '@/app/features/users/types/profile';
-import { upsertUserProfile } from '@/app/features/organization/services/profileService';
-import { updateUser } from '@/app/features/users/services/userService';
-import { RoleOptions } from '@/app/features/organization/pages/Organization/types';
-import { useNotify } from '@/app/hooks/useNotify';
-import {
-  getPreferredTimeZone,
-  resolveTimezoneFromCountry,
-  setPreferredTimeZone,
-} from '@/app/lib/timezone';
-import { useAuthStore } from '@/app/stores/authStore';
-import { getProfileForUserForPrimaryOrg } from '@/app/features/organization/services/teamService';
-import { useOrgStore } from '@/app/stores/orgStore';
-import '@/app/features/settings/styles/Settings.css';
-
-const ProfessionalFields = [
-  {
-    label: 'LinkedIn',
-    key: 'linkedin',
-    required: false,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'Medical license number',
-    key: 'medicalLicenseNumber',
-    required: false,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'Years of experience',
-    key: 'yearsOfExperience',
-    required: true,
-    editable: true,
-    type: 'number',
-  },
-  {
-    label: 'Specialisation',
-    key: 'specialization',
-    required: true,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'Qualification (MBBS, MD,etc.)',
-    key: 'qualification',
-    required: true,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'Biography or short description',
-    key: 'biography',
-    required: false,
-    editable: true,
-    type: 'text',
-  },
-];
-
-const AddressFields = [
-  {
-    label: 'Address line',
-    key: 'addressLine',
-    required: true,
-    editable: true,
-    type: 'googleAddress',
-  },
-  {
-    label: 'State / Province',
-    key: 'state',
-    required: true,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'City',
-    key: 'city',
-    required: true,
-    editable: true,
-    type: 'text',
-  },
-  {
-    label: 'Postal code',
-    key: 'postalCode',
-    required: true,
-    editable: true,
-    type: 'text',
-  },
-];
-
-const UserOrgProfileFields = [
-  { label: 'First name', key: 'given_name', required: true, editable: true, type: 'text' },
-  { label: 'Last name', key: 'family_name', required: true, editable: true, type: 'text' },
-  { label: 'Email address', key: 'email', required: true, editable: false, type: 'text' },
-  { label: 'Org name', key: 'name', required: false, editable: false, type: 'text' },
-  {
-    label: 'Role',
-    key: 'roleDisplay',
-    required: false,
-    editable: false,
-    type: 'select',
-    options: RoleOptions,
-  },
-  {
-    label: 'Employment type',
-    key: 'employmentType',
-    required: false,
-    editable: false,
-    type: 'select',
-    options: UserEmploymentTypeOptions,
-  },
-  { label: '', key: '_sep1', required: false, editable: false, type: 'separator' },
-  {
-    label: 'Gender',
-    key: 'gender',
-    required: false,
-    editable: true,
-    type: 'select',
-    options: UserGenderOptions,
-  },
-  {
-    label: 'Date of birth',
-    key: 'dateOfBirth',
-    required: true,
-    editable: true,
-    type: 'dateString',
-  },
-  { label: 'Phone number', key: 'phoneNumber', required: false, editable: true, type: 'text' },
-  { label: 'Country', key: 'country', required: false, editable: true, type: 'country' },
-];
-
-const OrgSection = () => {
-  const attributes = useAuthStore((s) => s.attributes);
-  const primaryOrgId = useOrgStore((s) => s.primaryOrgId);
-  const { org, membership } = usePrimaryOrgWithMembership();
-  const { availabilities } = usePrimaryAvailability();
-  const { notify } = useNotify();
-  const profile = usePrimaryOrgProfile();
-  const [availability, setAvailability] = useState<AvailabilityState>(() =>
-    daysOfWeek.reduce<AvailabilityState>((acc, day) => {
-      const isWeekday =
-        day === 'Monday' ||
-        day === 'Tuesday' ||
-        day === 'Wednesday' ||
-        day === 'Thursday' ||
-        day === 'Friday';
-
-      acc[day] = {
-        enabled: isWeekday,
-        intervals: [{ ...DEFAULT_INTERVAL }],
-      };
-      return acc;
-    }, {} as AvailabilityState)
-  );
-  const [isSavingAvailability, setIsSavingAvailability] = useState(false);
-  const userOrgProfileData = useMemo(
-    () => ({
-      given_name: attributes?.given_name ?? '',
-      family_name: attributes?.family_name ?? '',
-      email: attributes?.email ?? '',
-      name: org?.name ?? '',
-      roleDisplay: membership?.roleDisplay ?? '',
-      employmentType: profile?.personalDetails?.employmentType ?? '',
-      gender: profile?.personalDetails?.gender ?? '',
-      dateOfBirth: profile?.personalDetails?.dateOfBirth ?? '',
-      phoneNumber: profile?.personalDetails?.phoneNumber ?? '',
-      country: profile?.personalDetails?.address?.country ?? '',
-    }),
-    [attributes, org, membership, profile]
-  );
-
-  const addressData = useMemo(
-    () => ({
-      addressLine: profile?.personalDetails?.address?.addressLine ?? '',
-      state: profile?.personalDetails?.address?.state ?? '',
-      city: profile?.personalDetails?.address?.city ?? '',
-      postalCode: profile?.personalDetails?.address?.postalCode ?? '',
-    }),
-    [profile]
-  );
-
-  const professionalData = useMemo(
-    () => ({
-      linkedin: profile?.professionalDetails?.linkedin ?? '',
-      medicalLicenseNumber: profile?.professionalDetails?.medicalLicenseNumber ?? '',
-      yearsOfExperience: profile?.professionalDetails?.yearsOfExperience ?? '',
-      specialization: profile?.professionalDetails?.specialization ?? '',
-      qualification: profile?.professionalDetails?.qualification ?? '',
-      biography: profile?.professionalDetails?.biography ?? '',
-    }),
-    [profile]
-  );
-
-  // Render-phase adjustment: seed the editable availability from the store
-  // snapshot whenever a new snapshot arrives (org-level fallback only).
-  const [syncedAvailabilities, setSyncedAvailabilities] = useState<typeof availabilities | null>(
-    null
-  );
-  if (
-    availabilities &&
-    !membership?.practitionerReference &&
-    availabilities !== syncedAvailabilities
-  ) {
-    setSyncedAvailabilities(availabilities);
-    setAvailability(availabilities);
-  }
-
-  useEffect(() => {
-    const practitionerId = membership?.practitionerReference;
-    if (!practitionerId) return;
-    let cancelled = false;
-
-    (async () => {
-      const data = await getProfileForUserForPrimaryOrg(practitionerId);
-      if (cancelled) return;
-      const response = data as { baseAvailability?: unknown };
-      const baseAvailability = Array.isArray(response?.baseAvailability)
-        ? response.baseAvailability
-        : [];
-      setAvailability(convertFromGetApi(baseAvailability));
-    })().catch(() => undefined);
-
-    return () => {
-      cancelled = true;
-    };
-  }, [membership?.practitionerReference]);
-
-  const handleSaveAvailability = async () => {
-    if (isSavingAvailability) return;
-    try {
-      startTransition(() => setIsSavingAvailability(true));
-      await new Promise<void>((resolve) => {
-        setTimeout(resolve, 0);
-      });
-      const converted = convertAvailability(availability);
-      if (!hasAtLeastOneAvailability(converted)) {
-        console.log('No availability selected');
-        return;
-      }
-      if (membership?.practitionerReference) {
-        await upsertTeamAvailability(
-          {
-            _id: membership.id ?? membership.practitionerReference,
-            practionerId: membership.practitionerReference,
-            organisationId: primaryOrgId ?? String(org?._id ?? ''),
-          } as any,
-          converted,
-          null
-        );
-      } else {
-        await upsertAvailability(converted, null);
-      }
-      notify('success', {
-        title: 'Availability updated',
-        text: 'Availability have been updated successfully.',
-      });
-    } catch (error) {
-      console.log(error);
-      notify('error', {
-        title: 'Unable to update availability details',
-        text: 'Failed to update availability details. Please try again.',
-      });
-    } finally {
-      setIsSavingAvailability(false);
-    }
-  };
-
-  if (!attributes || !org || !membership) return null;
-
-  // Design subtitle for the availability editor: "<practitioner> · drives booking
-  // slots and the team planner", collapsing to the trailing clause when the signed-in
-  // user has no name on their auth attributes.
-  const practitionerName = `${attributes.given_name ?? ''} ${attributes.family_name ?? ''}`.trim();
-  const availabilitySubtitle = [practitionerName, 'drives booking slots and the team planner']
-    .filter((part) => part.length > 0)
-    .join(' · ');
-
-  const updateUserOrgProfileFields = async (values: any) => {
-    try {
-      await updateUser(values.given_name, values.family_name);
-      if (profile) {
-        const payload: UserProfile = {
-          ...profile,
-          _id: profile._id,
-          personalDetails: {
-            ...profile.personalDetails,
-            gender: values.gender as Gender,
-            dateOfBirth: values.dateOfBirth,
-            phoneNumber: values.phoneNumber,
-            address: {
-              ...profile.personalDetails?.address,
-              country: values.country,
-            },
-          },
-        };
-        await upsertUserProfile(payload);
-        const resolvedTimezone = resolveTimezoneFromCountry(values.country);
-        if (resolvedTimezone) setPreferredTimeZone(resolvedTimezone);
-      }
-      notify('success', {
-        title: 'Profile updated',
-        text: 'User profile has been updated successfully.',
-      });
-    } catch (error) {
-      console.log(error);
-      notify('error', {
-        title: 'Unable to update profile',
-        text: 'Failed to update user profile. Please try again.',
-      });
-    }
-  };
-
-  const updateAddressFields = async (values: any) => {
-    try {
-      if (!profile) return;
-      const payload: UserProfile = {
-        ...profile,
-        _id: profile?._id,
-        personalDetails: {
-          ...profile?.personalDetails,
-          address: {
-            ...profile?.personalDetails?.address,
-            addressLine: values.addressLine,
-            state: values.state,
-            city: values.city,
-            postalCode: values.postalCode,
-          },
-        },
-      };
-      await upsertUserProfile(payload);
-      notify('success', {
-        title: 'Address details updated',
-        text: 'Address details have been updated successfully.',
-      });
-    } catch (error) {
-      console.log(error);
-      notify('error', {
-        title: 'Unable to update address details',
-        text: 'Failed to update address details. Please try again.',
-      });
-    }
-  };
-
-  const updateProfessionalFields = async (values: any) => {
-    try {
-      if (!profile) return;
-      const payload: UserProfile = {
-        ...profile,
-        _id: profile?._id,
-        professionalDetails: {
-          ...profile?.professionalDetails,
-          linkedin: values.linkedin,
-          medicalLicenseNumber: values.medicalLicenseNumber,
-          specialization: values.specialization,
-          qualification: values.qualification,
-          biography: values.biography,
-          yearsOfExperience: values.yearsOfExperience,
-        },
-      };
-      await upsertUserProfile(payload);
-      notify('success', {
-        title: 'Professional details updated',
-        text: 'Professional details have been updated successfully.',
-      });
-    } catch (error) {
-      console.log(error);
-      notify('error', {
-        title: 'Unable to update professional details',
-        text: 'Failed to update professional details. Please try again.',
-      });
-    }
-  };
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div id="settings-user-profile" className="scroll-mt-24">
-        <ProfileCard
-          title="User profile"
-          fields={UserOrgProfileFields}
-          org={userOrgProfileData}
-          showProfileUser
-          onSave={updateUserOrgProfileFields}
-        />
-      </div>
-      <div
-        id="settings-availability"
-        className="scroll-mt-24 bg-[var(--screen)] border border-[var(--hairline)] rounded-[18px] shadow-[0_1px_2px_var(--sh03),0_8px_22px_var(--sh05)]"
-      >
-        <div className="px-5! pt-4! pb-3! border-b border-[var(--hairline)] flex items-center justify-between gap-3">
-          <div className="min-w-0">
-            <div className="text-[17px] font-bold tracking-[-0.02em] text-[var(--ink)]">
-              Availability &amp; consultation hours
-            </div>
-            <span className="yc-settings-editor-subtitle">{availabilitySubtitle}</span>
-          </div>
-        </div>
-        <div className="flex flex-col px-5! py-5!">
-          <Availability
-            availability={availability}
-            setAvailability={setAvailability}
-            twoColumnLayout
-          />
-        </div>
-        <div className="yc-settings-editor-footer">
-          <span className="yc-settings-tz-note">
-            <IoGlobeOutline size={14} aria-hidden="true" className="yc-settings-tz-note-icon" />
-            {getPreferredTimeZone()} · booking slots follow each service&apos;s duration
-          </span>
-          <Primary
-            href="#"
-            text={isSavingAvailability ? 'Saving...' : 'Save availability'}
-            onClick={handleSaveAvailability}
-            isDisabled={isSavingAvailability}
-          />
-        </div>
-      </div>
-      <ProfileCard
-        title="Address"
-        fields={AddressFields}
-        org={addressData}
-        onSave={updateAddressFields}
-      />
-      <ProfileCard
-        title="Professional details"
-        fields={ProfessionalFields}
-        org={professionalData}
-        onSave={updateProfessionalFields}
-      />
-    </div>
-  );
-};
-
-export default OrgSection;
